+++ src/test/java/com/intel/gkl/compression/InflaterUnitTest.java
@@ -1,8 +1,8 @@
 package com.intel.gkl.compression;
 
 import com.intel.gkl.IntelGKLUtils;
-import com.intel.gkl.testing_utils.CompressionDataProviders;
-import com.intel.gkl.testing_utils.TestingUtils;
+import com.intel.gkl.testingutils.CompressionDataProviders;
+import com.intel.gkl.testingutils.TestingUtils;
 import htsjdk.samtools.util.BlockCompressedInputStream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
