--- src/test/java/com/intel/gkl/testing_utils/TestingUtils.java
@@ -1,38 +1,0 @@
-package com.intel.gkl.testing_utils;
-
-import java.security.SecureRandom;
-import java.util.Arrays;
-
-public class TestingUtils {
-    public static final int SIZE_4MB = 4 *1024 * 1024;
-    public static final int MAX_ARRAY_SIZE = Integer.MAX_VALUE -2;
-
-    static SecureRandom rng = new SecureRandom();
-
-    public static void randomDNA(byte[] array) {
-        final byte[] DNA_CHARS = {'A', 'C', 'G', 'T'};
-
-        for (int i = 0; i < array.length; i++) {
-            array[i] = DNA_CHARS[rng.nextInt(DNA_CHARS.length)];
-        }
-    }
-
-    public static byte[] generateRandomDNAArray(int size){
-        byte[] dna = new byte[size];
-        randomDNA(dna);
-        return dna;
-    }
-    public static byte[] generateSimpleDNAArray(int size) {
-        byte[] dna = new byte[size];
-        Arrays.fill(dna, (byte) 'A');
-        return dna;
-    }
-    public static int getMaxInputSizeForGivenOutputSize(int outputBufferSize, int level, boolean nowrap){
-        if(level != 0){
-            return outputBufferSize;
-        }
-        return (int) (0.99 * outputBufferSize);
-    }
-
-
-}
